requirements.txt
python-telegram-bot==13.7
redis==5.0.1
Requests==2.31.0
urllib3==1.26.18
mysql-connector-python
flask
python-dotenv